src/e/forms/FormDialog.java: src/e/util/InstanceTracker.java: src/e/util/NumberDecoder.java: remove redundant casts spotted by Java 6 build 60's javac(1).

+++ salma-hayek/trunk/src/e/util/NumberDecoder.java
@@ -44,7 +44,7 @@ public class NumberDecoder {
         Long result;
         try {
             result = Long.valueOf(s.substring(index), radix);
-            result = negative ? Long.valueOf((long)-result.longValue()) : result;
+            result = negative ? Long.valueOf(-result.longValue()) : result;
             number = result.longValue();
             valid = true;
         } catch (NumberFormatException ex) {
